+++ examples/multi_thread.rs
@@ -1,7 +1,6 @@
 #[macro_use] extern crate logkit;
 
-#[tokio::main]
-async fn main() -> std::io::Result<()> {
+fn main() -> std::io::Result<()> {
     let mut logger = logkit::Logger::new(Some(&logkit::StdoutTarget));
     logger.mount(logkit::TimePlugin::from_millis());
     logger.mount(logkit::LevelPlugin);
@@ -15,12 +14,14 @@ async fn main() -> std::io::Result<()> {
     let mut handles = vec![];
 
     for i in 0..100 {
-        handles.push(tokio::task::spawn(async move {
+        handles.push(std::thread::spawn(move || {
             trace!("hello, this is a log with index {}", i);
         }));
     }
 
-    futures::future::join_all(handles).await;
+    for handle in handles {
+        handle.join().unwrap();
+    }
 
     Ok(())
 }