+++ src/plugin.rs
@@ -63,7 +63,7 @@ pub struct TimePlugin {
     pub local: bool,
 
     /// time format
-    pub format: time::format_description::OwnedFormatItem,
+    pub format: chrono::SecondsFormat,
 }
 
 impl TimePlugin {
@@ -73,7 +73,7 @@ impl TimePlugin {
     /// {"time":"2024-01-03T11:01:00+08:00"}
     /// ```
     pub fn from_secs() -> Self {
-        Self { local: true, format: time::format_description::parse_owned::<2>("[year]-[month]-[day]T[hour]:[minute]:[second][offset_hour sign:mandatory]:[offset_minute]").unwrap_or_else(|_| unreachable!()) }
+        Self { local: true, format: chrono::SecondsFormat::Secs }
     }
 
     /// Millisecond-level precision
@@ -82,7 +82,7 @@ impl TimePlugin {
     /// {"time":"2024-01-03T11:01:00.123+08:00"}
     /// ```
     pub fn from_millis() -> Self {
-        Self { local: true, format: time::format_description::parse_owned::<2>("[year]-[month]-[day]T[hour]:[minute]:[second].[subsecond digits:3][offset_hour sign:mandatory]:[offset_minute]").unwrap_or_else(|_| unreachable!()) }
+        Self { local: true, format: chrono::SecondsFormat::Millis }
     }
 
     /// Microsecond-level precision
@@ -91,7 +91,7 @@ impl TimePlugin {
     /// {"time":"2024-01-03T11:01:00.123456+08:00"}
     /// ```
     pub fn from_micros() -> Self {
-        Self { local: true, format: time::format_description::parse_owned::<2>("[year]-[month]-[day]T[hour]:[minute]:[second].[subsecond digits:6][offset_hour sign:mandatory]:[offset_minute]").unwrap_or_else(|_| unreachable!()) }
+        Self { local: true, format: chrono::SecondsFormat::Micros }
     }
 
     /// Nanosecond-level precision
@@ -100,7 +100,7 @@ impl TimePlugin {
     /// {"time":"2024-01-03T11:01:00.123456789+08:00"}
     /// ```
     pub fn from_nanos() -> Self {
-        Self { local: true, format: time::format_description::parse_owned::<2>("[year]-[month]-[day]T[hour]:[minute]:[second].[subsecond digits:9][offset_hour sign:mandatory]:[offset_minute]").unwrap_or_else(|_| unreachable!()) }
+        Self { local: true, format: chrono::SecondsFormat::Nanos }
     }
 
     /// Use UTC for formatting
@@ -118,11 +118,11 @@ impl Plugin for TimePlugin {
     #[inline]
     fn pre(&self, record: &mut Record) -> bool {
         let now = match self.local {
-            true => time::OffsetDateTime::now_local().unwrap_or(time::OffsetDateTime::now_utc()),
-            false => time::OffsetDateTime::now_utc(),
+            true => chrono::Local::now().to_rfc3339_opts(self.format, false),
+            false => chrono::Utc::now().to_rfc3339_opts(self.format, false),
         };
 
-        record.append("time", &now.format(&self.format).unwrap_or_default());
+        record.append("time", &now);
         true
     }
 }